Stage02_trace_analysis/requirements.txt
# HPC工作负载分析项目依赖包
# 基于Helios项目的分析需求

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 配置文件处理
PyYAML>=6.0

# 日期时间处理
python-dateutil>=2.8.0

# 科学计算
scipy>=1.7.0

# 统计分析
statsmodels>=0.13.0

# 可选：Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0

# 可选：进度条
tqdm>=4.60.0
